frozen-requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dotted==0.1.8
flatten-json==0.1.7
google-api-core==1.19.0
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gsheets==0.4.1
httplib2==0.18.1
idna==2.9
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
swagger-client==1.0.0
uritemplate==3.0.1
urllib3==1.25.9
